refactor: split tenant/security/isolation.go into 4 domain files (Issue #451)
isolation.go conflated four distinct security concerns in a single 1,794-line
file with 44 types and 89 methods, making the codebase harder to navigate and
reason about. Each domain — isolation rule management, vulnerability scanning,
zero-trust profiles, and adaptive controls — is a self-contained concern with
its own type hierarchy and logic.

Splitting along clean domain boundaries improves readability and allows
developers to find relevant code without scanning the entire file.

Changes:
- Extract vulnerability_manager.go (439 lines): VulnerabilityType, Vulnerability,
  RemediationPlan, VulnerabilityAssessment types + ScanForVulnerabilities,
  detectVulnerabilities, calculateRiskScore, GetVulnerabilities,
  UpdateVulnerabilityStatus, ExecuteRemediationPlan methods
- Extract zerotrust_profiles.go (494 lines): ZeroTrustProfile, ZeroTrustLevel,
  ZeroTrustDeviceProfile, BehavioralBaseline, AccessPattern types +
  InitializeZeroTrustProfile, UpdateTrustLevel, RecordDeviceFingerprint,
  EstablishBehavioralBaseline, EvaluateZeroTrustAccess + trust evaluation helpers
- Extract adaptive_controls.go (371 lines): AdaptiveSecurityControl, RiskLevel,
  AdaptationRule, AdaptationTriggerType, AdaptiveAuthConfig, ZeroTrustAccessEvent/
  Request/Response types + InitializeAdaptiveControls, triggerAdaptiveControls,
  executeAdaptationRule + action methods
- Trim isolation.go (511 lines): TenantIsolationEngine struct, IsolationRule CRUD,
  access validation, network/resource restriction helpers

All files remain in the security package; no public API surface changed.
All 26 tests in features/tenant/security pass (measured in test run).
Cross-platform compilation verified (linux/amd64, linux/arm64, darwin, windows).
Architecture compliance check passes (make check-architecture).

Fixes #451
