+++ features/tenant/security/adaptive_controls.go
@@ -1,0 +1,371 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Jordan Ritz
+package security
+
+import (
+	"context"
+	"fmt"
+	"time"
+)
+
+// AdaptiveAuthConfig defines adaptive authentication settings
+type AdaptiveAuthConfig struct {
+	MFARequired           bool                   `json:"mfa_required"`
+	RiskBasedMFA          bool                   `json:"risk_based_mfa"`
+	RiskThreshold         float64                `json:"risk_threshold"`
+	AdditionalFactors     []AuthenticationFactor `json:"additional_factors"`
+	ContinuousAuth        bool                   `json:"continuous_auth"`
+	SessionTimeout        time.Duration          `json:"session_timeout"`
+	ReauthenticationRules []ReauthenticationRule `json:"reauthentication_rules"`
+}
+
+// AuthenticationFactor defines authentication factors
+type AuthenticationFactor string
+
+const (
+	AuthFactorPassword       AuthenticationFactor = "password"
+	AuthFactorTOTP           AuthenticationFactor = "totp"
+	AuthFactorSMS            AuthenticationFactor = "sms"
+	AuthFactorBiometric      AuthenticationFactor = "biometric"
+	AuthFactorHardwareToken  AuthenticationFactor = "hardware_token"
+	AuthFactorDeviceLocation AuthenticationFactor = "device_location"
+)
+
+// ReauthenticationRule defines when reauthentication is required
+type ReauthenticationRule struct {
+	Trigger     string        `json:"trigger"`
+	Condition   string        `json:"condition"`
+	GracePeriod time.Duration `json:"grace_period"`
+}
+
+// ContextualControl defines context-based security controls
+type ContextualControl struct {
+	ControlID  string                 `json:"control_id"`
+	Type       ContextualControlType  `json:"type"`
+	Condition  string                 `json:"condition"`
+	Action     string                 `json:"action"`
+	Parameters map[string]interface{} `json:"parameters"`
+	Enabled    bool                   `json:"enabled"`
+	Priority   int                    `json:"priority"`
+}
+
+// ContextualControlType defines types of contextual controls
+type ContextualControlType string
+
+const (
+	ContextualControlTypeLocation ContextualControlType = "location"
+	ContextualControlTypeTime     ContextualControlType = "time"
+	ContextualControlTypeDevice   ContextualControlType = "device"
+	ContextualControlTypeRisk     ContextualControlType = "risk"
+	ContextualControlTypeData     ContextualControlType = "data"
+	ContextualControlTypeBehavior ContextualControlType = "behavior"
+)
+
+// ZeroTrustAccessEvent represents an access event for behavioral analysis
+type ZeroTrustAccessEvent struct {
+	EventID    string                 `json:"event_id"`
+	TenantID   string                 `json:"tenant_id"`
+	UserID     string                 `json:"user_id"`
+	DeviceID   string                 `json:"device_id"`
+	ResourceID string                 `json:"resource_id"`
+	Action     string                 `json:"action"`
+	Timestamp  time.Time              `json:"timestamp"`
+	Location   string                 `json:"location,omitempty"`
+	Success    bool                   `json:"success"`
+	Duration   time.Duration          `json:"duration,omitempty"`
+	DataVolume float64                `json:"data_volume,omitempty"`
+	Context    map[string]interface{} `json:"context,omitempty"`
+}
+
+// ZeroTrustAccessRequest represents a zero-trust access request
+type ZeroTrustAccessRequest struct {
+	RequestID    string                 `json:"request_id"`
+	TenantID     string                 `json:"tenant_id"`
+	UserID       string                 `json:"user_id"`
+	DeviceID     string                 `json:"device_id"`
+	ResourceID   string                 `json:"resource_id"`
+	ResourceType string                 `json:"resource_type"`
+	Action       string                 `json:"action"`
+	Timestamp    time.Time              `json:"timestamp"`
+	Location     string                 `json:"location,omitempty"`
+	Context      map[string]interface{} `json:"context,omitempty"`
+}
+
+// ZeroTrustAccessResponse represents the response to a zero-trust access request
+type ZeroTrustAccessResponse struct {
+	RequestID    string                 `json:"request_id"`
+	TenantID     string                 `json:"tenant_id"`
+	Granted      bool                   `json:"granted"`
+	TrustLevel   ZeroTrustLevel         `json:"trust_level"`
+	RiskScore    float64                `json:"risk_score"`
+	Reason       string                 `json:"reason"`
+	RequiredAuth []AuthenticationFactor `json:"required_auth"`
+	EvaluatedAt  time.Time              `json:"evaluated_at"`
+	Constraints  []string               `json:"constraints,omitempty"`
+	SessionData  map[string]interface{} `json:"session_data,omitempty"`
+}
+
+// AdaptiveSecurityControl defines adaptive security controls that adjust based on risk
+type AdaptiveSecurityControl struct {
+	TenantID           string              `json:"tenant_id"`
+	CurrentRiskLevel   RiskLevel           `json:"current_risk_level"`
+	AdaptationRules    []AdaptationRule    `json:"adaptation_rules"`
+	SecurityPosture    SecurityPosture     `json:"security_posture"`
+	AutomatedResponses []AutomatedResponse `json:"automated_responses"`
+	EscalationPolicy   *EscalationPolicy   `json:"escalation_policy"`
+	LastAdaptation     time.Time           `json:"last_adaptation"`
+	AdaptationHistory  []AdaptationEvent   `json:"adaptation_history"`
+}
+
+// RiskLevel defines risk assessment levels
+type RiskLevel string
+
+const (
+	RiskLevelLow      RiskLevel = "low"
+	RiskLevelMedium   RiskLevel = "medium"
+	RiskLevelHigh     RiskLevel = "high"
+	RiskLevelCritical RiskLevel = "critical"
+)
+
+// AdaptationRule defines how security controls adapt to risk changes
+type AdaptationRule struct {
+	RuleID      string                 `json:"rule_id"`
+	TriggerType AdaptationTriggerType  `json:"trigger_type"`
+	Threshold   float64                `json:"threshold"`
+	Action      AdaptationAction       `json:"action"`
+	Parameters  map[string]interface{} `json:"parameters"`
+	Enabled     bool                   `json:"enabled"`
+}
+
+// AdaptationTriggerType defines what triggers adaptation
+type AdaptationTriggerType string
+
+const (
+	AdaptationTriggerRiskIncrease     AdaptationTriggerType = "risk_increase"
+	AdaptationTriggerThreatDetection  AdaptationTriggerType = "threat_detection"
+	AdaptationTriggerAnomalyDetected  AdaptationTriggerType = "anomaly_detected"
+	AdaptationTriggerComplianceChange AdaptationTriggerType = "compliance_change"
+)
+
+// AdaptationAction defines actions taken during adaptation
+type AdaptationAction string
+
+const (
+	AdaptationActionTightenControls    AdaptationAction = "tighten_controls"
+	AdaptationActionRequireMFA         AdaptationAction = "require_mfa"
+	AdaptationActionLimitAccess        AdaptationAction = "limit_access"
+	AdaptationActionIncreaseMonitoring AdaptationAction = "increase_monitoring"
+	AdaptationActionIsolateSession     AdaptationAction = "isolate_session"
+	AdaptationActionEscalateAlert      AdaptationAction = "escalate_alert"
+)
+
+// SecurityPosture defines the current security stance
+type SecurityPosture struct {
+	PostureLevel    SecurityPostureLevel `json:"posture_level"`
+	Controls        []string             `json:"active_controls"`
+	Restrictions    []string             `json:"active_restrictions"`
+	MonitoringLevel string               `json:"monitoring_level"`
+	LastUpdate      time.Time            `json:"last_update"`
+}
+
+// SecurityPostureLevel defines security posture levels
+type SecurityPostureLevel string
+
+const (
+	SecurityPostureRelaxed    SecurityPostureLevel = "relaxed"
+	SecurityPostureNormal     SecurityPostureLevel = "normal"
+	SecurityPostureHeightened SecurityPostureLevel = "heightened"
+	SecurityPostureRestricted SecurityPostureLevel = "restricted"
+	SecurityPostureLockdown   SecurityPostureLevel = "lockdown"
+)
+
+// AutomatedResponse defines automated security responses
+type AutomatedResponse struct {
+	ResponseID  string                 `json:"response_id"`
+	TriggerType string                 `json:"trigger_type"`
+	Action      string                 `json:"action"`
+	Parameters  map[string]interface{} `json:"parameters"`
+	Enabled     bool                   `json:"enabled"`
+}
+
+// EscalationPolicy defines escalation procedures
+type EscalationPolicy struct {
+	Levels   []EscalationLevel `json:"levels"`
+	Enabled  bool              `json:"enabled"`
+	MaxLevel int               `json:"max_level"`
+}
+
+// EscalationLevel defines an escalation level
+type EscalationLevel struct {
+	Level     int           `json:"level"`
+	Threshold float64       `json:"threshold"`
+	Actions   []string      `json:"actions"`
+	Timeout   time.Duration `json:"timeout"`
+	Contacts  []string      `json:"contacts"`
+}
+
+// AdaptationEvent records adaptation history
+type AdaptationEvent struct {
+	EventID   string                 `json:"event_id"`
+	Timestamp time.Time              `json:"timestamp"`
+	Trigger   string                 `json:"trigger"`
+	Action    string                 `json:"action"`
+	Result    string                 `json:"result"`
+	Context   map[string]interface{} `json:"context"`
+}
+
+// InitializeAdaptiveControls initializes adaptive security controls for a tenant
+func (tie *TenantIsolationEngine) InitializeAdaptiveControls(ctx context.Context, tenantID string) (*AdaptiveSecurityControl, error) {
+	tie.mutex.Lock()
+	defer tie.mutex.Unlock()
+
+	controls := &AdaptiveSecurityControl{
+		TenantID:         tenantID,
+		CurrentRiskLevel: RiskLevelMedium,
+		AdaptationRules:  tie.createDefaultAdaptationRules(),
+		SecurityPosture: SecurityPosture{
+			PostureLevel:    SecurityPostureNormal,
+			Controls:        []string{"basic_authentication", "access_logging"},
+			Restrictions:    []string{},
+			MonitoringLevel: "standard",
+			LastUpdate:      time.Now(),
+		},
+		AutomatedResponses: tie.createDefaultAutomatedResponses(),
+		EscalationPolicy:   tie.createDefaultEscalationPolicy(),
+		LastAdaptation:     time.Now(),
+		AdaptationHistory:  []AdaptationEvent{},
+	}
+
+	tie.adaptiveControls[tenantID] = controls
+	return controls, nil
+}
+
+// triggerAdaptiveControls triggers adaptive security controls based on risk changes
+func (tie *TenantIsolationEngine) triggerAdaptiveControls(ctx context.Context, tenantID string, trigger AdaptationTriggerType) error {
+	controls, exists := tie.adaptiveControls[tenantID]
+	if !exists {
+		// Initialize if not exists
+		var err error
+		controls, err = tie.InitializeAdaptiveControls(ctx, tenantID)
+		if err != nil {
+			return err
+		}
+	}
+
+	// Find applicable adaptation rules
+	for _, rule := range controls.AdaptationRules {
+		if rule.Enabled && rule.TriggerType == trigger {
+			err := tie.executeAdaptationRule(ctx, tenantID, rule)
+			if err != nil {
+				continue // Try next rule
+			}
+
+			// Record adaptation event
+			event := AdaptationEvent{
+				EventID:   fmt.Sprintf("adapt-%d", time.Now().UnixNano()),
+				Timestamp: time.Now(),
+				Trigger:   string(trigger),
+				Action:    string(rule.Action),
+				Result:    "success",
+				Context:   rule.Parameters,
+			}
+			controls.AdaptationHistory = append(controls.AdaptationHistory, event)
+			controls.LastAdaptation = time.Now()
+		}
+	}
+
+	return nil
+}
+
+func (tie *TenantIsolationEngine) createDefaultAdaptationRules() []AdaptationRule {
+	return []AdaptationRule{
+		{
+			RuleID:      "risk_increase_mfa",
+			TriggerType: AdaptationTriggerRiskIncrease,
+			Threshold:   0.8,
+			Action:      AdaptationActionRequireMFA,
+			Parameters:  map[string]interface{}{"factors": []string{"totp", "biometric"}},
+			Enabled:     true,
+		},
+		{
+			RuleID:      "threat_detection_isolate",
+			TriggerType: AdaptationTriggerThreatDetection,
+			Threshold:   0.9,
+			Action:      AdaptationActionIsolateSession,
+			Parameters:  map[string]interface{}{"duration": "1h"},
+			Enabled:     true,
+		},
+	}
+}
+
+func (tie *TenantIsolationEngine) createDefaultAutomatedResponses() []AutomatedResponse {
+	return []AutomatedResponse{
+		{
+			ResponseID:  "high_risk_alert",
+			TriggerType: "high_risk_detected",
+			Action:      "send_alert",
+			Parameters:  map[string]interface{}{"severity": "high"},
+			Enabled:     true,
+		},
+	}
+}
+
+func (tie *TenantIsolationEngine) createDefaultEscalationPolicy() *EscalationPolicy {
+	return &EscalationPolicy{
+		Levels: []EscalationLevel{
+			{Level: 1, Threshold: 0.7, Actions: []string{"alert"}, Timeout: time.Minute * 5},
+			{Level: 2, Threshold: 0.8, Actions: []string{"alert", "limit_access"}, Timeout: time.Minute * 10},
+			{Level: 3, Threshold: 0.9, Actions: []string{"alert", "isolate", "notify_admin"}, Timeout: time.Minute * 15},
+		},
+		Enabled:  true,
+		MaxLevel: 3,
+	}
+}
+
+func (tie *TenantIsolationEngine) executeAdaptationRule(ctx context.Context, tenantID string, rule AdaptationRule) error {
+	switch rule.Action {
+	case AdaptationActionTightenControls:
+		return tie.tightenSecurityControls(ctx, tenantID)
+	case AdaptationActionRequireMFA:
+		return tie.enableAdditionalMFA(ctx, tenantID, rule.Parameters)
+	case AdaptationActionLimitAccess:
+		return tie.limitTenantAccess(ctx, tenantID, rule.Parameters)
+	case AdaptationActionIncreaseMonitoring:
+		return tie.increaseMonitoring(ctx, tenantID)
+	case AdaptationActionIsolateSession:
+		return tie.isolateActiveSessions(ctx, tenantID)
+	case AdaptationActionEscalateAlert:
+		return tie.escalateSecurityAlert(ctx, tenantID, rule.Parameters)
+	}
+	return nil
+}
+
+func (tie *TenantIsolationEngine) tightenSecurityControls(ctx context.Context, tenantID string) error {
+	// Implementation would tighten various security controls
+	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("controls-%s", tenantID), "tighten_controls")
+}
+
+func (tie *TenantIsolationEngine) enableAdditionalMFA(ctx context.Context, tenantID string, params map[string]interface{}) error {
+	// Implementation would enable additional MFA requirements
+	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("mfa-%s", tenantID), "enable_mfa")
+}
+
+func (tie *TenantIsolationEngine) limitTenantAccess(ctx context.Context, tenantID string, params map[string]interface{}) error {
+	// Implementation would limit access permissions
+	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("access-%s", tenantID), "limit_access")
+}
+
+func (tie *TenantIsolationEngine) increaseMonitoring(ctx context.Context, tenantID string) error {
+	// Implementation would increase monitoring levels
+	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("monitor-%s", tenantID), "increase_monitoring")
+}
+
+func (tie *TenantIsolationEngine) isolateActiveSessions(ctx context.Context, tenantID string) error {
+	// Implementation would isolate active sessions
+	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("session-%s", tenantID), "isolate_sessions")
+}
+
+func (tie *TenantIsolationEngine) escalateSecurityAlert(ctx context.Context, tenantID string, params map[string]interface{}) error {
+	// Implementation would escalate security alerts
+	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("alert-%s", tenantID), "escalate_alert")
+}
