+++ features/tenant/security/isolation.go
@@ -100,338 +100,27 @@ const (
 	ComplianceLevelCustom  ComplianceLevel = "custom"
 )
 
-// ZeroTrustProfile defines zero-trust security profile for a tenant
-type ZeroTrustProfile struct {
-	TenantID               string                             `json:"tenant_id"`
-	TrustLevel             ZeroTrustLevel                     `json:"trust_level"`
-	DeviceFingerprints     map[string]*ZeroTrustDeviceProfile `json:"device_fingerprints"`
-	BehavioralBaseline     *BehavioralBaseline                `json:"behavioral_baseline"`
-	AccessPatterns         []AccessPattern                    `json:"access_patterns"`
-	RiskScore              float64                            `json:"risk_score"`
-	ContinuousVerification bool                               `json:"continuous_verification"`
-	AdaptiveAuthentication *AdaptiveAuthConfig                `json:"adaptive_authentication"`
-	ContextualControls     []ContextualControl                `json:"contextual_controls"`
-	LastUpdated            time.Time                          `json:"last_updated"`
+// TenantAccessRequest represents a request to access tenant resources
+type TenantAccessRequest struct {
+	SubjectID       string                       `json:"subject_id"`
+	SubjectTenantID string                       `json:"subject_tenant_id"`
+	TargetTenantID  string                       `json:"target_tenant_id"`
+	ResourceID      string                       `json:"resource_id"`
+	AccessLevel     CrossTenantLevel             `json:"access_level"`
+	Context         map[string]string            `json:"context"`
+	AuthContext     *common.AuthorizationContext `json:"auth_context,omitempty"`
 }
 
-// ZeroTrustLevel defines trust levels in zero-trust model
-type ZeroTrustLevel string
-
-const (
-	ZeroTrustLevelUntrusted ZeroTrustLevel = "untrusted"
-	ZeroTrustLevelLow       ZeroTrustLevel = "low"
-	ZeroTrustLevelMedium    ZeroTrustLevel = "medium"
-	ZeroTrustLevelHigh      ZeroTrustLevel = "high"
-	ZeroTrustLevelVerified  ZeroTrustLevel = "verified"
-)
-
-// ZeroTrustDeviceProfile stores device fingerprinting information for zero-trust
-type ZeroTrustDeviceProfile struct {
-	DeviceID           string                 `json:"device_id"`
-	DeviceType         string                 `json:"device_type"`
-	OSVersion          string                 `json:"os_version"`
-	BrowserFingerprint string                 `json:"browser_fingerprint,omitempty"`
-	NetworkFingerprint string                 `json:"network_fingerprint"`
-	TrustScore         float64                `json:"trust_score"`
-	LastSeen           time.Time              `json:"last_seen"`
-	Attributes         map[string]interface{} `json:"attributes"`
-	RiskIndicators     []string               `json:"risk_indicators"`
-}
-
-// BehavioralBaseline defines normal behavior patterns for a tenant
-type BehavioralBaseline struct {
-	TypicalAccessHours   []TimeWindow      `json:"typical_access_hours"`
-	CommonLocations      []GeographicZone  `json:"common_locations"`
-	UsualResources       []string          `json:"usual_resources"`
-	AverageSessionLength time.Duration     `json:"average_session_length"`
-	NormalDataVolume     DataVolumePattern `json:"normal_data_volume"`
-	EstablishedAt        time.Time         `json:"established_at"`
-	ConfidenceLevel      float64           `json:"confidence_level"`
-}
-
-// TimeWindow defines a time range for access patterns
-type TimeWindow struct {
-	StartHour int    `json:"start_hour"`
-	EndHour   int    `json:"end_hour"`
-	DayOfWeek string `json:"day_of_week,omitempty"`
-	Timezone  string `json:"timezone"`
-}
-
-// GeographicZone defines geographic areas for access patterns
-type GeographicZone struct {
-	Country string  `json:"country"`
-	Region  string  `json:"region,omitempty"`
-	City    string  `json:"city,omitempty"`
-	Radius  float64 `json:"radius,omitempty"` // km radius for area
-}
-
-// DataVolumePattern defines normal data access patterns
-type DataVolumePattern struct {
-	AverageReadMB  float64 `json:"average_read_mb"`
-	AverageWriteMB float64 `json:"average_write_mb"`
-	PeakReadMB     float64 `json:"peak_read_mb"`
-	PeakWriteMB    float64 `json:"peak_write_mb"`
-}
-
-// AccessPattern defines observed access patterns
-type AccessPattern struct {
-	PatternID  string                 `json:"pattern_id"`
-	Type       AccessPatternType      `json:"type"`
-	Frequency  int                    `json:"frequency"`
-	LastSeen   time.Time              `json:"last_seen"`
-	Confidence float64                `json:"confidence"`
-	Context    map[string]interface{} `json:"context"`
-	RiskLevel  string                 `json:"risk_level"`
-}
-
-// AccessPatternType defines types of access patterns
-type AccessPatternType string
-
-const (
-	AccessPatternTypeNormal     AccessPatternType = "normal"
-	AccessPatternTypeSuspicious AccessPatternType = "suspicious"
-	AccessPatternTypeAnomaly    AccessPatternType = "anomaly"
-	AccessPatternTypeBaseline   AccessPatternType = "baseline"
-)
-
-// AdaptiveAuthConfig defines adaptive authentication settings
-type AdaptiveAuthConfig struct {
-	MFARequired           bool                   `json:"mfa_required"`
-	RiskBasedMFA          bool                   `json:"risk_based_mfa"`
-	RiskThreshold         float64                `json:"risk_threshold"`
-	AdditionalFactors     []AuthenticationFactor `json:"additional_factors"`
-	ContinuousAuth        bool                   `json:"continuous_auth"`
-	SessionTimeout        time.Duration          `json:"session_timeout"`
-	ReauthenticationRules []ReauthenticationRule `json:"reauthentication_rules"`
-}
-
-// AuthenticationFactor defines authentication factors
-type AuthenticationFactor string
-
-const (
-	AuthFactorPassword       AuthenticationFactor = "password"
-	AuthFactorTOTP           AuthenticationFactor = "totp"
-	AuthFactorSMS            AuthenticationFactor = "sms"
-	AuthFactorBiometric      AuthenticationFactor = "biometric"
-	AuthFactorHardwareToken  AuthenticationFactor = "hardware_token"
-	AuthFactorDeviceLocation AuthenticationFactor = "device_location"
-)
-
-// ReauthenticationRule defines when reauthentication is required
-type ReauthenticationRule struct {
-	Trigger     string        `json:"trigger"`
-	Condition   string        `json:"condition"`
-	GracePeriod time.Duration `json:"grace_period"`
-}
-
-// ContextualControl defines context-based security controls
-type ContextualControl struct {
-	ControlID  string                 `json:"control_id"`
-	Type       ContextualControlType  `json:"type"`
-	Condition  string                 `json:"condition"`
-	Action     string                 `json:"action"`
-	Parameters map[string]interface{} `json:"parameters"`
-	Enabled    bool                   `json:"enabled"`
-	Priority   int                    `json:"priority"`
-}
-
-// ContextualControlType defines types of contextual controls
-type ContextualControlType string
-
-const (
-	ContextualControlTypeLocation ContextualControlType = "location"
-	ContextualControlTypeTime     ContextualControlType = "time"
-	ContextualControlTypeDevice   ContextualControlType = "device"
-	ContextualControlTypeRisk     ContextualControlType = "risk"
-	ContextualControlTypeData     ContextualControlType = "data"
-	ContextualControlTypeBehavior ContextualControlType = "behavior"
-)
-
-// AdaptiveSecurityControl defines adaptive security controls that adjust based on risk
-type AdaptiveSecurityControl struct {
-	TenantID           string              `json:"tenant_id"`
-	CurrentRiskLevel   RiskLevel           `json:"current_risk_level"`
-	AdaptationRules    []AdaptationRule    `json:"adaptation_rules"`
-	SecurityPosture    SecurityPosture     `json:"security_posture"`
-	AutomatedResponses []AutomatedResponse `json:"automated_responses"`
-	EscalationPolicy   *EscalationPolicy   `json:"escalation_policy"`
-	LastAdaptation     time.Time           `json:"last_adaptation"`
-	AdaptationHistory  []AdaptationEvent   `json:"adaptation_history"`
-}
-
-// RiskLevel defines risk assessment levels
-type RiskLevel string
-
-const (
-	RiskLevelLow      RiskLevel = "low"
-	RiskLevelMedium   RiskLevel = "medium"
-	RiskLevelHigh     RiskLevel = "high"
-	RiskLevelCritical RiskLevel = "critical"
-)
-
-// AdaptationRule defines how security controls adapt to risk changes
-type AdaptationRule struct {
-	RuleID      string                 `json:"rule_id"`
-	TriggerType AdaptationTriggerType  `json:"trigger_type"`
-	Threshold   float64                `json:"threshold"`
-	Action      AdaptationAction       `json:"action"`
-	Parameters  map[string]interface{} `json:"parameters"`
-	Enabled     bool                   `json:"enabled"`
-}
-
-// AdaptationTriggerType defines what triggers adaptation
-type AdaptationTriggerType string
-
-const (
-	AdaptationTriggerRiskIncrease     AdaptationTriggerType = "risk_increase"
-	AdaptationTriggerThreatDetection  AdaptationTriggerType = "threat_detection"
-	AdaptationTriggerAnomalyDetected  AdaptationTriggerType = "anomaly_detected"
-	AdaptationTriggerComplianceChange AdaptationTriggerType = "compliance_change"
-)
-
-// AdaptationAction defines actions taken during adaptation
-type AdaptationAction string
-
-const (
-	AdaptationActionTightenControls    AdaptationAction = "tighten_controls"
-	AdaptationActionRequireMFA         AdaptationAction = "require_mfa"
-	AdaptationActionLimitAccess        AdaptationAction = "limit_access"
-	AdaptationActionIncreaseMonitoring AdaptationAction = "increase_monitoring"
-	AdaptationActionIsolateSession     AdaptationAction = "isolate_session"
-	AdaptationActionEscalateAlert      AdaptationAction = "escalate_alert"
-)
-
-// SecurityPosture defines the current security stance
-type SecurityPosture struct {
-	PostureLevel    SecurityPostureLevel `json:"posture_level"`
-	Controls        []string             `json:"active_controls"`
-	Restrictions    []string             `json:"active_restrictions"`
-	MonitoringLevel string               `json:"monitoring_level"`
-	LastUpdate      time.Time            `json:"last_update"`
-}
-
-// SecurityPostureLevel defines security posture levels
-type SecurityPostureLevel string
-
-const (
-	SecurityPostureRelaxed    SecurityPostureLevel = "relaxed"
-	SecurityPostureNormal     SecurityPostureLevel = "normal"
-	SecurityPostureHeightened SecurityPostureLevel = "heightened"
-	SecurityPostureRestricted SecurityPostureLevel = "restricted"
-	SecurityPostureLockdown   SecurityPostureLevel = "lockdown"
-)
-
-// AutomatedResponse defines automated security responses
-type AutomatedResponse struct {
-	ResponseID  string                 `json:"response_id"`
-	TriggerType string                 `json:"trigger_type"`
-	Action      string                 `json:"action"`
-	Parameters  map[string]interface{} `json:"parameters"`
-	Enabled     bool                   `json:"enabled"`
-}
-
-// EscalationPolicy defines escalation procedures
-type EscalationPolicy struct {
-	Levels   []EscalationLevel `json:"levels"`
-	Enabled  bool              `json:"enabled"`
-	MaxLevel int               `json:"max_level"`
-}
-
-// EscalationLevel defines an escalation level
-type EscalationLevel struct {
-	Level     int           `json:"level"`
-	Threshold float64       `json:"threshold"`
-	Actions   []string      `json:"actions"`
-	Timeout   time.Duration `json:"timeout"`
-	Contacts  []string      `json:"contacts"`
-}
-
-// AdaptationEvent records adaptation history
-type AdaptationEvent struct {
-	EventID   string                 `json:"event_id"`
-	Timestamp time.Time              `json:"timestamp"`
-	Trigger   string                 `json:"trigger"`
-	Action    string                 `json:"action"`
-	Result    string                 `json:"result"`
-	Context   map[string]interface{} `json:"context"`
-}
-
-// VulnerabilityType defines types of security vulnerabilities
-type VulnerabilityType string
-
-const (
-	VulnerabilityMissingPatches      VulnerabilityType = "missing_patches"
-	VulnerabilityWeakCredentials     VulnerabilityType = "weak_credentials"
-	VulnerabilityOpenPorts           VulnerabilityType = "open_ports"
-	VulnerabilityOutdatedSoftware    VulnerabilityType = "outdated_software"
-	VulnerabilityMisconfiguration    VulnerabilityType = "misconfiguration"
-	VulnerabilityPrivilegeEscalation VulnerabilityType = "privilege_escalation"
-	VulnerabilityDataExposure        VulnerabilityType = "data_exposure"
-)
-
-// VulnerabilitySeverity defines vulnerability severity levels
-type VulnerabilitySeverity string
-
-const (
-	VulnerabilitySeverityLow      VulnerabilitySeverity = "low"
-	VulnerabilitySeverityMedium   VulnerabilitySeverity = "medium"
-	VulnerabilitySeverityHigh     VulnerabilitySeverity = "high"
-	VulnerabilitySeverityCritical VulnerabilitySeverity = "critical"
-)
-
-// RemediationAction defines automated remediation actions
-type RemediationAction string
-
-const (
-	RemediationActionQuarantine    RemediationAction = "quarantine"
-	RemediationActionPatch         RemediationAction = "patch"
-	RemediationActionDisable       RemediationAction = "disable"
-	RemediationActionAlert         RemediationAction = "alert"
-	RemediationActionBlock         RemediationAction = "block"
-	RemediationActionRotateSecrets RemediationAction = "rotate_secrets"
-	RemediationActionIsolate       RemediationAction = "isolate"
-)
-
-// Vulnerability represents a detected security vulnerability
-type Vulnerability struct {
-	ID            string                `json:"id"`
-	TenantID      string                `json:"tenant_id"`
-	Type          VulnerabilityType     `json:"type"`
-	Severity      VulnerabilitySeverity `json:"severity"`
-	Title         string                `json:"title"`
-	Description   string                `json:"description"`
-	AffectedAsset string                `json:"affected_asset"`
-	CVEID         string                `json:"cve_id,omitempty"`
-	CVSSScore     float64               `json:"cvss_score,omitempty"`
-	DetectedAt    time.Time             `json:"detected_at"`
-	UpdatedAt     time.Time             `json:"updated_at"`
-	Status        string                `json:"status"` // "open", "remediated", "mitigated", "accepted"
-	Remediation   *RemediationPlan      `json:"remediation,omitempty"`
-}
-
-// RemediationPlan defines how to remediate a vulnerability
-type RemediationPlan struct {
-	Actions       []RemediationAction `json:"actions"`
-	AutoRemediate bool                `json:"auto_remediate"`
-	ManualSteps   []string            `json:"manual_steps"`
-	EstimatedTime time.Duration       `json:"estimated_time"`
-	Priority      int                 `json:"priority"` // 1-10, where 1 is highest priority
-	Dependencies  []string            `json:"dependencies"`
-	RollbackPlan  []string            `json:"rollback_plan"`
-	CreatedAt     time.Time           `json:"created_at"`
-	ScheduledAt   *time.Time          `json:"scheduled_at,omitempty"`
-	CompletedAt   *time.Time          `json:"completed_at,omitempty"`
-}
-
-// VulnerabilityAssessment contains the result of vulnerability scanning
-type VulnerabilityAssessment struct {
-	TenantID         string          `json:"tenant_id"`
-	ScanID           string          `json:"scan_id"`
-	ScanTime         time.Time       `json:"scan_time"`
-	Vulnerabilities  []Vulnerability `json:"vulnerabilities"`
-	RiskScore        float64         `json:"risk_score"`
-	ComplianceStatus map[string]bool `json:"compliance_status"`
-	Recommendations  []string        `json:"recommendations"`
+// TenantAccessResponse represents the response to a tenant access request
+type TenantAccessResponse struct {
+	Granted        bool             `json:"granted"`
+	TenantID       string           `json:"tenant_id"`
+	SubjectID      string           `json:"subject_id"`
+	ResourceID     string           `json:"resource_id"`
+	RequestedLevel CrossTenantLevel `json:"requested_level"`
+	Reason         string           `json:"reason"`
+	EffectiveRule  *IsolationRule   `json:"effective_rule,omitempty"`
+	ValidationTime time.Time        `json:"validation_time"`
 }
 
 // NewTenantIsolationEngine creates a new tenant isolation engine
@@ -607,6 +296,55 @@ func (tie *TenantIsolationEngine) ValidateTenantAccess(ctx context.Context, requ
 	return response, nil
 }
 
+// ValidateTenantResourceAccess provides a simplified interface for validating tenant access
+// This method is used by the TenantSecretManager and other components that need
+// simple resource access validation within a tenant context
+func (tie *TenantIsolationEngine) ValidateTenantResourceAccess(tenantID, resourceType, permission string) bool {
+	tie.mutex.RLock()
+	defer tie.mutex.RUnlock()
+
+	// Get isolation rule for the tenant
+	rule, exists := tie.isolationRules[tenantID]
+	if !exists {
+		// Use default rule if none exists
+		rule = tie.getDefaultIsolationRule(tenantID)
+	}
+
+	// For secrets management, check resource isolation settings
+	if resourceType == "secrets" {
+		// Check if the resource type is explicitly restricted
+		for _, restricted := range rule.ResourceIsolation.RestrictedResources {
+			if restricted == resourceType {
+				return false
+			}
+		}
+
+		// If isolated storage is required, we allow operations within the tenant
+		// This means each tenant has their own isolated secret storage
+		if rule.ResourceIsolation.IsolatedStorage {
+			return true
+		}
+
+		// Check if resource sharing is allowed
+		if rule.ResourceIsolation.AllowResourceSharing {
+			return true
+		}
+
+		// Default: allow access within tenant for secret operations
+		return true
+	}
+
+	// For other resource types, apply basic validation
+	// Check if the resource type is in the restricted list
+	for _, restricted := range rule.ResourceIsolation.RestrictedResources {
+		if restricted == resourceType {
+			return false
+		}
+	}
+
+	return true
+}
+
 // validateIsolationRule validates an isolation rule for correctness
 func (tie *TenantIsolationEngine) validateIsolationRule(rule *IsolationRule) error {
 	if rule.TenantID == "" {
@@ -770,1025 +508,4 @@ func (tie *TenantIsolationEngine) getDefaultIsolationRule(tenantID string) *Isol
 		CreatedAt:       time.Now(),
 		UpdatedAt:       time.Now(),
 	}
-}
-
-// TenantAccessRequest represents a request to access tenant resources
-type TenantAccessRequest struct {
-	SubjectID       string                       `json:"subject_id"`
-	SubjectTenantID string                       `json:"subject_tenant_id"`
-	TargetTenantID  string                       `json:"target_tenant_id"`
-	ResourceID      string                       `json:"resource_id"`
-	AccessLevel     CrossTenantLevel             `json:"access_level"`
-	Context         map[string]string            `json:"context"`
-	AuthContext     *common.AuthorizationContext `json:"auth_context,omitempty"`
-}
-
-// TenantAccessResponse represents the response to a tenant access request
-type TenantAccessResponse struct {
-	Granted        bool             `json:"granted"`
-	TenantID       string           `json:"tenant_id"`
-	SubjectID      string           `json:"subject_id"`
-	ResourceID     string           `json:"resource_id"`
-	RequestedLevel CrossTenantLevel `json:"requested_level"`
-	Reason         string           `json:"reason"`
-	EffectiveRule  *IsolationRule   `json:"effective_rule,omitempty"`
-	ValidationTime time.Time        `json:"validation_time"`
-}
-
-// ValidateTenantResourceAccess provides a simplified interface for validating tenant access
-// This method is used by the TenantSecretManager and other components that need
-// simple resource access validation within a tenant context
-func (tie *TenantIsolationEngine) ValidateTenantResourceAccess(tenantID, resourceType, permission string) bool {
-	tie.mutex.RLock()
-	defer tie.mutex.RUnlock()
-
-	// Get isolation rule for the tenant
-	rule, exists := tie.isolationRules[tenantID]
-	if !exists {
-		// Use default rule if none exists
-		rule = tie.getDefaultIsolationRule(tenantID)
-	}
-
-	// For secrets management, check resource isolation settings
-	if resourceType == "secrets" {
-		// Check if the resource type is explicitly restricted
-		for _, restricted := range rule.ResourceIsolation.RestrictedResources {
-			if restricted == resourceType {
-				return false
-			}
-		}
-
-		// If isolated storage is required, we allow operations within the tenant
-		// This means each tenant has their own isolated secret storage
-		if rule.ResourceIsolation.IsolatedStorage {
-			return true
-		}
-
-		// Check if resource sharing is allowed
-		if rule.ResourceIsolation.AllowResourceSharing {
-			return true
-		}
-
-		// Default: allow access within tenant for secret operations
-		return true
-	}
-
-	// For other resource types, apply basic validation
-	// Check if the resource type is in the restricted list
-	for _, restricted := range rule.ResourceIsolation.RestrictedResources {
-		if restricted == resourceType {
-			return false
-		}
-	}
-
-	return true
-}
-
-// ScanForVulnerabilities performs a comprehensive vulnerability assessment for a tenant
-func (tie *TenantIsolationEngine) ScanForVulnerabilities(ctx context.Context, tenantID string) (*VulnerabilityAssessment, error) {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	// Validate tenant exists
-	_, err := tie.tenantManager.GetTenant(ctx, tenantID)
-	if err != nil {
-		return nil, fmt.Errorf("tenant not found: %w", err)
-	}
-
-	assessment := &VulnerabilityAssessment{
-		TenantID:         tenantID,
-		ScanID:           fmt.Sprintf("scan-%s-%d", tenantID, time.Now().Unix()),
-		ScanTime:         time.Now(),
-		Vulnerabilities:  []Vulnerability{},
-		ComplianceStatus: make(map[string]bool),
-		Recommendations:  []string{},
-	}
-
-	// Get isolation rule for compliance requirements
-	rule, _ := tie.GetIsolationRule(ctx, tenantID)
-
-	// Simulate vulnerability detection (in production, this would integrate with actual scanners)
-	vulnerabilities := tie.detectVulnerabilities(tenantID, rule)
-	assessment.Vulnerabilities = vulnerabilities
-
-	// Calculate risk score based on vulnerabilities
-	assessment.RiskScore = tie.calculateRiskScore(vulnerabilities)
-
-	// Check compliance status
-	assessment.ComplianceStatus = tie.checkComplianceStatus(rule, vulnerabilities)
-
-	// Generate recommendations
-	assessment.Recommendations = tie.generateRecommendations(vulnerabilities, rule)
-
-	// Store vulnerabilities for tracking
-	tie.vulnerabilities[tenantID] = vulnerabilities
-
-	// Auto-create remediation plans for critical vulnerabilities
-	for _, vuln := range vulnerabilities {
-		if vuln.Severity == VulnerabilitySeverityCritical || vuln.Severity == VulnerabilitySeverityHigh {
-			plan := tie.createRemediationPlan(vuln, rule)
-			tie.remediationPlans[vuln.ID] = plan
-		}
-	}
-
-	return assessment, nil
-}
-
-// detectVulnerabilities simulates vulnerability detection
-func (tie *TenantIsolationEngine) detectVulnerabilities(tenantID string, rule *IsolationRule) []Vulnerability {
-	vulnerabilities := []Vulnerability{}
-	now := time.Now()
-
-	// Check for weak credential policies
-	if !rule.NetworkIsolation.RequireMTLS {
-		vulnerabilities = append(vulnerabilities, Vulnerability{
-			ID:            fmt.Sprintf("vuln-%s-weak-auth", tenantID),
-			TenantID:      tenantID,
-			Type:          VulnerabilityWeakCredentials,
-			Severity:      VulnerabilitySeverityHigh,
-			Title:         "Weak Authentication Configuration",
-			Description:   "Mutual TLS is not required, potentially allowing weaker authentication methods",
-			AffectedAsset: "authentication_system",
-			DetectedAt:    now,
-			UpdatedAt:     now,
-			Status:        "open",
-		})
-	}
-
-	// Check for inadequate encryption
-	if !rule.DataResidency.RequireEncryption || rule.DataResidency.EncryptionLevel == "standard" {
-		severity := VulnerabilitySeverityMedium
-		if !rule.DataResidency.RequireEncryption {
-			severity = VulnerabilitySeverityHigh
-		}
-		vulnerabilities = append(vulnerabilities, Vulnerability{
-			ID:            fmt.Sprintf("vuln-%s-encryption", tenantID),
-			TenantID:      tenantID,
-			Type:          VulnerabilityDataExposure,
-			Severity:      severity,
-			Title:         "Inadequate Data Encryption",
-			Description:   "Data encryption requirements are insufficient for security compliance",
-			AffectedAsset: "data_storage",
-			DetectedAt:    now,
-			UpdatedAt:     now,
-			Status:        "open",
-		})
-	}
-
-	// Check for overly permissive cross-tenant access
-	if rule.CrossTenantAccess.AllowCrossTenantAccess && !rule.CrossTenantAccess.RequireApproval {
-		vulnerabilities = append(vulnerabilities, Vulnerability{
-			ID:            fmt.Sprintf("vuln-%s-cross-tenant", tenantID),
-			TenantID:      tenantID,
-			Type:          VulnerabilityPrivilegeEscalation,
-			Severity:      VulnerabilitySeverityMedium,
-			Title:         "Overly Permissive Cross-Tenant Access",
-			Description:   "Cross-tenant access is allowed without requiring approval, increasing attack surface",
-			AffectedAsset: "access_control",
-			DetectedAt:    now,
-			UpdatedAt:     now,
-			Status:        "open",
-		})
-	}
-
-	// Check for insufficient resource isolation
-	if !rule.ResourceIsolation.IsolatedStorage {
-		vulnerabilities = append(vulnerabilities, Vulnerability{
-			ID:            fmt.Sprintf("vuln-%s-isolation", tenantID),
-			TenantID:      tenantID,
-			Type:          VulnerabilityMisconfiguration,
-			Severity:      VulnerabilitySeverityHigh,
-			Title:         "Insufficient Resource Isolation",
-			Description:   "Tenant resources are not properly isolated, potentially allowing data leakage",
-			AffectedAsset: "resource_isolation",
-			DetectedAt:    now,
-			UpdatedAt:     now,
-			Status:        "open",
-		})
-	}
-
-	return vulnerabilities
-}
-
-// calculateRiskScore calculates an overall risk score based on vulnerabilities
-func (tie *TenantIsolationEngine) calculateRiskScore(vulnerabilities []Vulnerability) float64 {
-	if len(vulnerabilities) == 0 {
-		return 0.0
-	}
-
-	var totalScore float64
-	weights := map[VulnerabilitySeverity]float64{
-		VulnerabilitySeverityLow:      1.0,
-		VulnerabilitySeverityMedium:   3.0,
-		VulnerabilitySeverityHigh:     7.0,
-		VulnerabilitySeverityCritical: 10.0,
-	}
-
-	for _, vuln := range vulnerabilities {
-		if weight, exists := weights[vuln.Severity]; exists {
-			totalScore += weight
-		}
-	}
-
-	// Normalize to 0-100 scale
-	maxPossibleScore := float64(len(vulnerabilities)) * 10.0
-	if maxPossibleScore > 0 {
-		return (totalScore / maxPossibleScore) * 100.0
-	}
-
-	return 0.0
-}
-
-// checkComplianceStatus checks compliance against various standards
-func (tie *TenantIsolationEngine) checkComplianceStatus(rule *IsolationRule, vulnerabilities []Vulnerability) map[string]bool {
-	status := make(map[string]bool)
-
-	// Check basic compliance requirements
-	status["basic"] = true
-	for _, vuln := range vulnerabilities {
-		if vuln.Severity == VulnerabilitySeverityCritical {
-			status["basic"] = false
-			break
-		}
-	}
-
-	// Check specific compliance standards based on tenant's compliance level
-	switch rule.ComplianceLevel {
-	case ComplianceLevelHIPAA:
-		status["hipaa"] = rule.DataResidency.RequireEncryption &&
-			rule.DataResidency.EncryptionLevel != "standard" &&
-			rule.ResourceIsolation.IsolatedStorage
-	case ComplianceLevelPCIDSS:
-		status["pci_dss"] = rule.NetworkIsolation.RequireMTLS &&
-			rule.DataResidency.RequireEncryption &&
-			rule.ResourceIsolation.IsolatedStorage
-	case ComplianceLevelFedRAMP:
-		status["fedramp"] = rule.DataResidency.EncryptionLevel == "fips" &&
-			rule.NetworkIsolation.RequireMTLS &&
-			!rule.CrossTenantAccess.AllowCrossTenantAccess
-	}
-
-	return status
-}
-
-// generateRecommendations generates security recommendations based on vulnerabilities
-func (tie *TenantIsolationEngine) generateRecommendations(vulnerabilities []Vulnerability, rule *IsolationRule) []string {
-	recommendations := []string{}
-
-	for _, vuln := range vulnerabilities {
-		switch vuln.Type {
-		case VulnerabilityWeakCredentials:
-			recommendations = append(recommendations, "Enable mutual TLS authentication")
-			recommendations = append(recommendations, "Implement multi-factor authentication")
-		case VulnerabilityDataExposure:
-			recommendations = append(recommendations, "Upgrade to high-level encryption")
-			recommendations = append(recommendations, "Enable encryption at rest and in transit")
-		case VulnerabilityPrivilegeEscalation:
-			recommendations = append(recommendations, "Require approval for cross-tenant access")
-			recommendations = append(recommendations, "Implement least privilege access controls")
-		case VulnerabilityMisconfiguration:
-			recommendations = append(recommendations, "Enable isolated storage for tenant resources")
-			recommendations = append(recommendations, "Review and update security configurations")
-		}
-	}
-
-	return recommendations
-}
-
-// createRemediationPlan creates an automated remediation plan for a vulnerability
-func (tie *TenantIsolationEngine) createRemediationPlan(vuln Vulnerability, rule *IsolationRule) *RemediationPlan {
-	plan := &RemediationPlan{
-		Actions:       []RemediationAction{},
-		AutoRemediate: false,
-		ManualSteps:   []string{},
-		EstimatedTime: time.Hour,
-		Priority:      5,
-		Dependencies:  []string{},
-		RollbackPlan:  []string{},
-		CreatedAt:     time.Now(),
-	}
-
-	switch vuln.Type {
-	case VulnerabilityWeakCredentials:
-		plan.Actions = []RemediationAction{RemediationActionAlert}
-		plan.ManualSteps = []string{
-			"Enable mutual TLS in network isolation settings",
-			"Update authentication configuration",
-			"Test authentication with new settings",
-		}
-		plan.Priority = 2
-		plan.EstimatedTime = time.Hour * 2
-
-	case VulnerabilityDataExposure:
-		plan.Actions = []RemediationAction{RemediationActionAlert}
-		plan.ManualSteps = []string{
-			"Upgrade encryption level to 'high' or 'fips'",
-			"Enable encryption requirements",
-			"Verify data encryption across all systems",
-		}
-		plan.Priority = 1
-		plan.EstimatedTime = time.Hour * 4
-
-	case VulnerabilityPrivilegeEscalation:
-		plan.Actions = []RemediationAction{RemediationActionAlert}
-		plan.ManualSteps = []string{
-			"Enable approval requirement for cross-tenant access",
-			"Review existing cross-tenant permissions",
-			"Update access control policies",
-		}
-		plan.Priority = 3
-		plan.EstimatedTime = time.Hour * 1
-
-	case VulnerabilityMisconfiguration:
-		plan.Actions = []RemediationAction{RemediationActionAlert}
-		plan.ManualSteps = []string{
-			"Enable isolated storage in resource isolation settings",
-			"Migrate shared resources to isolated storage",
-			"Verify resource isolation boundaries",
-		}
-		plan.Priority = 2
-		plan.EstimatedTime = time.Hour * 6
-	}
-
-	if vuln.Severity == VulnerabilitySeverityCritical {
-		plan.Priority = 1
-		plan.Actions = append(plan.Actions, RemediationActionIsolate)
-	}
-
-	return plan
-}
-
-// GetVulnerabilities returns vulnerabilities for a tenant
-func (tie *TenantIsolationEngine) GetVulnerabilities(ctx context.Context, tenantID string) ([]Vulnerability, error) {
-	tie.mutex.RLock()
-	defer tie.mutex.RUnlock()
-
-	vulnerabilities, exists := tie.vulnerabilities[tenantID]
-	if !exists {
-		return []Vulnerability{}, nil
-	}
-
-	return vulnerabilities, nil
-}
-
-// UpdateVulnerabilityStatus updates the status of a vulnerability
-func (tie *TenantIsolationEngine) UpdateVulnerabilityStatus(ctx context.Context, vulnerabilityID, status string) error {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	// Find and update the vulnerability
-	for tenantID, vulnerabilities := range tie.vulnerabilities {
-		for i, vuln := range vulnerabilities {
-			if vuln.ID == vulnerabilityID {
-				vulnerabilities[i].Status = status
-				vulnerabilities[i].UpdatedAt = time.Now()
-
-				// If remediated, update remediation plan
-				if status == "remediated" {
-					if plan, exists := tie.remediationPlans[vulnerabilityID]; exists {
-						now := time.Now()
-						plan.CompletedAt = &now
-					}
-				}
-
-				// Audit the status change
-				return tie.auditLogger.LogVulnerabilityStatusChange(ctx, vulnerabilityID, tenantID, status)
-			}
-		}
-	}
-
-	return fmt.Errorf("vulnerability not found: %s", vulnerabilityID)
-}
-
-// ExecuteRemediationPlan executes an automated remediation plan
-func (tie *TenantIsolationEngine) ExecuteRemediationPlan(ctx context.Context, vulnerabilityID string) error {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	plan, exists := tie.remediationPlans[vulnerabilityID]
-	if !exists {
-		return fmt.Errorf("remediation plan not found for vulnerability: %s", vulnerabilityID)
-	}
-
-	if !plan.AutoRemediate {
-		return fmt.Errorf("remediation plan requires manual execution")
-	}
-
-	// Execute automated actions
-	for _, action := range plan.Actions {
-		switch action {
-		case RemediationActionAlert:
-			// Send alert to administrators
-			_ = tie.auditLogger.LogRemediationAction(ctx, vulnerabilityID, string(action))
-		case RemediationActionIsolate:
-			// Temporarily isolate affected resources
-			_ = tie.auditLogger.LogRemediationAction(ctx, vulnerabilityID, string(action))
-		case RemediationActionBlock:
-			// Block suspicious access
-			_ = tie.auditLogger.LogRemediationAction(ctx, vulnerabilityID, string(action))
-		}
-	}
-
-	// Mark as scheduled for execution
-	now := time.Now()
-	plan.ScheduledAt = &now
-
-	return nil
-}
-
-// Enhanced Zero-Trust Security Model Methods
-
-// InitializeZeroTrustProfile initializes a zero-trust profile for a tenant
-func (tie *TenantIsolationEngine) InitializeZeroTrustProfile(ctx context.Context, tenantID string) (*ZeroTrustProfile, error) {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	// Validate tenant exists
-	_, err := tie.tenantManager.GetTenant(ctx, tenantID)
-	if err != nil {
-		return nil, fmt.Errorf("tenant not found: %w", err)
-	}
-
-	profile := &ZeroTrustProfile{
-		TenantID:               tenantID,
-		TrustLevel:             ZeroTrustLevelUntrusted,
-		DeviceFingerprints:     make(map[string]*ZeroTrustDeviceProfile),
-		BehavioralBaseline:     tie.createDefaultBehavioralBaseline(),
-		AccessPatterns:         []AccessPattern{},
-		RiskScore:              50.0, // Start with medium risk
-		ContinuousVerification: true,
-		AdaptiveAuthentication: tie.createDefaultAdaptiveAuthConfig(),
-		ContextualControls:     tie.createDefaultContextualControls(),
-		LastUpdated:            time.Now(),
-	}
-
-	tie.zeroTrustProfiles[tenantID] = profile
-	return profile, nil
-}
-
-// UpdateTrustLevel updates the trust level for a tenant based on behavior
-func (tie *TenantIsolationEngine) UpdateTrustLevel(ctx context.Context, tenantID string, newLevel ZeroTrustLevel, reason string) error {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	profile, exists := tie.zeroTrustProfiles[tenantID]
-	if !exists {
-		return fmt.Errorf("zero-trust profile not found for tenant: %s", tenantID)
-	}
-
-	oldLevel := profile.TrustLevel
-	profile.TrustLevel = newLevel
-	profile.LastUpdated = time.Now()
-
-	// Update risk score based on trust level
-	profile.RiskScore = tie.calculateRiskScoreFromTrustLevel(newLevel)
-
-	// Trigger adaptive controls if trust level decreased
-	if tie.shouldTriggerAdaptation(oldLevel, newLevel) {
-		err := tie.triggerAdaptiveControls(ctx, tenantID, AdaptationTriggerRiskIncrease)
-		if err != nil {
-			// Log error but don't fail the trust level update
-			_ = tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("trust-level-%s", tenantID), "adaptation_failed")
-		}
-	}
-
-	// Audit the trust level change
-	return tie.auditLogger.LogVulnerabilityStatusChange(ctx, fmt.Sprintf("trust-level-%s", tenantID), tenantID, string(newLevel))
-}
-
-// RecordDeviceFingerprint records device fingerprinting information
-func (tie *TenantIsolationEngine) RecordDeviceFingerprint(ctx context.Context, tenantID string, device *ZeroTrustDeviceProfile) error {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	profile, exists := tie.zeroTrustProfiles[tenantID]
-	if !exists {
-		// Initialize profile if it doesn't exist
-		var err error
-		profile, err = tie.InitializeZeroTrustProfile(ctx, tenantID)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Update device profile
-	device.LastSeen = time.Now()
-	profile.DeviceFingerprints[device.DeviceID] = device
-	profile.LastUpdated = time.Now()
-
-	// Calculate device trust score
-	device.TrustScore = tie.calculateDeviceTrustScore(device)
-
-	// Update overall risk score based on device risk
-	tie.updateRiskScoreFromDevice(profile, device)
-
-	return nil
-}
-
-// EstablishBehavioralBaseline establishes behavioral baseline for a tenant
-func (tie *TenantIsolationEngine) EstablishBehavioralBaseline(ctx context.Context, tenantID string, accessEvents []ZeroTrustAccessEvent) error {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	profile, exists := tie.zeroTrustProfiles[tenantID]
-	if !exists {
-		return fmt.Errorf("zero-trust profile not found for tenant: %s", tenantID)
-	}
-
-	baseline := tie.analyzeBehavioralPatterns(accessEvents)
-	baseline.EstablishedAt = time.Now()
-	baseline.ConfidenceLevel = tie.calculateBaselineConfidence(accessEvents)
-
-	profile.BehavioralBaseline = baseline
-	profile.LastUpdated = time.Now()
-
-	return nil
-}
-
-// EvaluateZeroTrustAccess evaluates access request against zero-trust principles
-func (tie *TenantIsolationEngine) EvaluateZeroTrustAccess(ctx context.Context, request *ZeroTrustAccessRequest) (*ZeroTrustAccessResponse, error) {
-	tie.mutex.RLock()
-	defer tie.mutex.RUnlock()
-
-	profile, exists := tie.zeroTrustProfiles[request.TenantID]
-	if !exists {
-		return &ZeroTrustAccessResponse{
-			Granted:      false,
-			TrustLevel:   ZeroTrustLevelUntrusted,
-			RiskScore:    100.0,
-			Reason:       "No zero-trust profile established",
-			RequiredAuth: []AuthenticationFactor{AuthFactorPassword, AuthFactorTOTP},
-		}, nil
-	}
-
-	response := &ZeroTrustAccessResponse{
-		TenantID:    request.TenantID,
-		RequestID:   request.RequestID,
-		EvaluatedAt: time.Now(),
-		TrustLevel:  profile.TrustLevel,
-		RiskScore:   profile.RiskScore,
-	}
-
-	// Evaluate device trust
-	deviceTrust := tie.evaluateDeviceTrust(profile, request.DeviceID)
-
-	// Evaluate behavioral patterns
-	behaviorTrust := tie.evaluateBehavioralTrust(profile, request)
-
-	// Evaluate contextual factors
-	contextTrust := tie.evaluateContextualTrust(profile, request)
-
-	// Calculate overall trust score
-	overallTrust := (deviceTrust + behaviorTrust + contextTrust) / 3.0
-	response.RiskScore = 100.0 - (overallTrust * 100.0)
-
-	// Determine access decision
-	response.Granted = tie.determineZeroTrustAccess(profile, overallTrust, request)
-
-	if response.Granted {
-		response.Reason = "Access granted based on zero-trust evaluation"
-		response.RequiredAuth = tie.determineRequiredAuthentication(profile, overallTrust)
-	} else {
-		response.Reason = "Access denied - insufficient trust level"
-		response.RequiredAuth = []AuthenticationFactor{AuthFactorPassword, AuthFactorTOTP, AuthFactorBiometric}
-	}
-
-	return response, nil
-}
-
-// InitializeAdaptiveControls initializes adaptive security controls for a tenant
-func (tie *TenantIsolationEngine) InitializeAdaptiveControls(ctx context.Context, tenantID string) (*AdaptiveSecurityControl, error) {
-	tie.mutex.Lock()
-	defer tie.mutex.Unlock()
-
-	controls := &AdaptiveSecurityControl{
-		TenantID:         tenantID,
-		CurrentRiskLevel: RiskLevelMedium,
-		AdaptationRules:  tie.createDefaultAdaptationRules(),
-		SecurityPosture: SecurityPosture{
-			PostureLevel:    SecurityPostureNormal,
-			Controls:        []string{"basic_authentication", "access_logging"},
-			Restrictions:    []string{},
-			MonitoringLevel: "standard",
-			LastUpdate:      time.Now(),
-		},
-		AutomatedResponses: tie.createDefaultAutomatedResponses(),
-		EscalationPolicy:   tie.createDefaultEscalationPolicy(),
-		LastAdaptation:     time.Now(),
-		AdaptationHistory:  []AdaptationEvent{},
-	}
-
-	tie.adaptiveControls[tenantID] = controls
-	return controls, nil
-}
-
-// TriggerAdaptiveControls triggers adaptive security controls based on risk changes
-func (tie *TenantIsolationEngine) triggerAdaptiveControls(ctx context.Context, tenantID string, trigger AdaptationTriggerType) error {
-	controls, exists := tie.adaptiveControls[tenantID]
-	if !exists {
-		// Initialize if not exists
-		var err error
-		controls, err = tie.InitializeAdaptiveControls(ctx, tenantID)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Find applicable adaptation rules
-	for _, rule := range controls.AdaptationRules {
-		if rule.Enabled && rule.TriggerType == trigger {
-			err := tie.executeAdaptationRule(ctx, tenantID, rule)
-			if err != nil {
-				continue // Try next rule
-			}
-
-			// Record adaptation event
-			event := AdaptationEvent{
-				EventID:   fmt.Sprintf("adapt-%d", time.Now().UnixNano()),
-				Timestamp: time.Now(),
-				Trigger:   string(trigger),
-				Action:    string(rule.Action),
-				Result:    "success",
-				Context:   rule.Parameters,
-			}
-			controls.AdaptationHistory = append(controls.AdaptationHistory, event)
-			controls.LastAdaptation = time.Now()
-		}
-	}
-
-	return nil
-}
-
-// Helper methods for zero-trust implementation
-
-func (tie *TenantIsolationEngine) createDefaultBehavioralBaseline() *BehavioralBaseline {
-	return &BehavioralBaseline{
-		TypicalAccessHours: []TimeWindow{
-			{StartHour: 9, EndHour: 17, DayOfWeek: "weekday", Timezone: "UTC"},
-		},
-		CommonLocations:      []GeographicZone{},
-		UsualResources:       []string{},
-		AverageSessionLength: time.Hour,
-		NormalDataVolume: DataVolumePattern{
-			AverageReadMB:  10.0,
-			AverageWriteMB: 5.0,
-			PeakReadMB:     50.0,
-			PeakWriteMB:    25.0,
-		},
-		EstablishedAt:   time.Now(),
-		ConfidenceLevel: 0.0,
-	}
-}
-
-func (tie *TenantIsolationEngine) createDefaultAdaptiveAuthConfig() *AdaptiveAuthConfig {
-	return &AdaptiveAuthConfig{
-		MFARequired:       true,
-		RiskBasedMFA:      true,
-		RiskThreshold:     0.7,
-		AdditionalFactors: []AuthenticationFactor{AuthFactorTOTP},
-		ContinuousAuth:    false,
-		SessionTimeout:    time.Hour * 8,
-		ReauthenticationRules: []ReauthenticationRule{
-			{Trigger: "high_risk_detected", Condition: "risk_score > 0.8", GracePeriod: time.Minute * 5},
-		},
-	}
-}
-
-func (tie *TenantIsolationEngine) createDefaultContextualControls() []ContextualControl {
-	return []ContextualControl{
-		{
-			ControlID:  "location_control",
-			Type:       ContextualControlTypeLocation,
-			Condition:  "unknown_location",
-			Action:     "require_additional_auth",
-			Parameters: map[string]interface{}{"auth_factor": "totp"},
-			Enabled:    true,
-			Priority:   1,
-		},
-		{
-			ControlID:  "time_control",
-			Type:       ContextualControlTypeTime,
-			Condition:  "outside_business_hours",
-			Action:     "increase_monitoring",
-			Parameters: map[string]interface{}{"monitoring_level": "high"},
-			Enabled:    true,
-			Priority:   2,
-		},
-	}
-}
-
-func (tie *TenantIsolationEngine) createDefaultAdaptationRules() []AdaptationRule {
-	return []AdaptationRule{
-		{
-			RuleID:      "risk_increase_mfa",
-			TriggerType: AdaptationTriggerRiskIncrease,
-			Threshold:   0.8,
-			Action:      AdaptationActionRequireMFA,
-			Parameters:  map[string]interface{}{"factors": []string{"totp", "biometric"}},
-			Enabled:     true,
-		},
-		{
-			RuleID:      "threat_detection_isolate",
-			TriggerType: AdaptationTriggerThreatDetection,
-			Threshold:   0.9,
-			Action:      AdaptationActionIsolateSession,
-			Parameters:  map[string]interface{}{"duration": "1h"},
-			Enabled:     true,
-		},
-	}
-}
-
-func (tie *TenantIsolationEngine) createDefaultAutomatedResponses() []AutomatedResponse {
-	return []AutomatedResponse{
-		{
-			ResponseID:  "high_risk_alert",
-			TriggerType: "high_risk_detected",
-			Action:      "send_alert",
-			Parameters:  map[string]interface{}{"severity": "high"},
-			Enabled:     true,
-		},
-	}
-}
-
-func (tie *TenantIsolationEngine) createDefaultEscalationPolicy() *EscalationPolicy {
-	return &EscalationPolicy{
-		Levels: []EscalationLevel{
-			{Level: 1, Threshold: 0.7, Actions: []string{"alert"}, Timeout: time.Minute * 5},
-			{Level: 2, Threshold: 0.8, Actions: []string{"alert", "limit_access"}, Timeout: time.Minute * 10},
-			{Level: 3, Threshold: 0.9, Actions: []string{"alert", "isolate", "notify_admin"}, Timeout: time.Minute * 15},
-		},
-		Enabled:  true,
-		MaxLevel: 3,
-	}
-}
-
-func (tie *TenantIsolationEngine) calculateRiskScoreFromTrustLevel(level ZeroTrustLevel) float64 {
-	switch level {
-	case ZeroTrustLevelVerified:
-		return 10.0
-	case ZeroTrustLevelHigh:
-		return 25.0
-	case ZeroTrustLevelMedium:
-		return 50.0
-	case ZeroTrustLevelLow:
-		return 75.0
-	case ZeroTrustLevelUntrusted:
-		return 90.0
-	default:
-		return 50.0
-	}
-}
-
-func (tie *TenantIsolationEngine) shouldTriggerAdaptation(oldLevel, newLevel ZeroTrustLevel) bool {
-	// Trigger adaptation if trust level decreased
-	levels := map[ZeroTrustLevel]int{
-		ZeroTrustLevelUntrusted: 0,
-		ZeroTrustLevelLow:       1,
-		ZeroTrustLevelMedium:    2,
-		ZeroTrustLevelHigh:      3,
-		ZeroTrustLevelVerified:  4,
-	}
-	return levels[newLevel] < levels[oldLevel]
-}
-
-func (tie *TenantIsolationEngine) calculateDeviceTrustScore(device *ZeroTrustDeviceProfile) float64 {
-	score := 1.0
-
-	// Penalize for risk indicators
-	score -= float64(len(device.RiskIndicators)) * 0.1
-	if score < 0 {
-		score = 0
-	}
-
-	// Factor in device type (mobile devices might be less trusted)
-	if device.DeviceType == "mobile" {
-		score *= 0.9
-	}
-
-	return score
-}
-
-func (tie *TenantIsolationEngine) updateRiskScoreFromDevice(profile *ZeroTrustProfile, device *ZeroTrustDeviceProfile) {
-	// Simple risk calculation - in production this would be more sophisticated
-	if device.TrustScore < 0.5 {
-		profile.RiskScore = profile.RiskScore * 1.2 // Increase risk
-		if profile.RiskScore > 100 {
-			profile.RiskScore = 100
-		}
-	}
-}
-
-func (tie *TenantIsolationEngine) analyzeBehavioralPatterns(events []ZeroTrustAccessEvent) *BehavioralBaseline {
-	// Simplified baseline analysis - in production this would use ML
-	baseline := tie.createDefaultBehavioralBaseline()
-
-	if len(events) > 0 {
-		// Analyze typical access hours
-		hourCounts := make(map[int]int)
-		for _, event := range events {
-			hour := event.Timestamp.Hour()
-			hourCounts[hour]++
-		}
-
-		// Find most common hours (simplified)
-		maxCount := 0
-		var startHour, endHour int
-		for hour, count := range hourCounts {
-			if count > maxCount {
-				maxCount = count
-				startHour = hour
-				endHour = hour + 8 // Assume 8-hour work day
-			}
-		}
-
-		baseline.TypicalAccessHours[0].StartHour = startHour
-		baseline.TypicalAccessHours[0].EndHour = endHour
-	}
-
-	return baseline
-}
-
-func (tie *TenantIsolationEngine) calculateBaselineConfidence(events []ZeroTrustAccessEvent) float64 {
-	// Simple confidence calculation based on number of events
-	if len(events) < 10 {
-		return 0.1
-	} else if len(events) < 50 {
-		return 0.5
-	} else if len(events) < 100 {
-		return 0.8
-	}
-	return 0.9
-}
-
-func (tie *TenantIsolationEngine) evaluateDeviceTrust(profile *ZeroTrustProfile, deviceID string) float64 {
-	device, exists := profile.DeviceFingerprints[deviceID]
-	if !exists {
-		return 0.1 // Very low trust for unknown devices
-	}
-	return device.TrustScore
-}
-
-func (tie *TenantIsolationEngine) evaluateBehavioralTrust(profile *ZeroTrustProfile, request *ZeroTrustAccessRequest) float64 {
-	if profile.BehavioralBaseline == nil {
-		return 0.5 // Neutral trust if no baseline
-	}
-
-	// Check if access is within typical hours
-	currentHour := request.Timestamp.Hour()
-	for _, window := range profile.BehavioralBaseline.TypicalAccessHours {
-		if currentHour >= window.StartHour && currentHour <= window.EndHour {
-			return 0.8 // High trust for typical hours
-		}
-	}
-
-	return 0.3 // Lower trust for unusual hours
-}
-
-func (tie *TenantIsolationEngine) evaluateContextualTrust(profile *ZeroTrustProfile, request *ZeroTrustAccessRequest) float64 {
-	trust := 0.5 // Start with neutral trust
-
-	// Evaluate each contextual control
-	for _, control := range profile.ContextualControls {
-		if control.Enabled {
-			switch control.Type {
-			case ContextualControlTypeLocation:
-				// Check if location is known/trusted
-				if request.Location != "" {
-					trust += 0.1 // Slight increase for known location
-				}
-			case ContextualControlTypeRisk:
-				// Factor in current risk level
-				if profile.RiskScore < 30 {
-					trust += 0.2
-				} else if profile.RiskScore > 70 {
-					trust -= 0.2
-				}
-			}
-		}
-	}
-
-	if trust > 1.0 {
-		trust = 1.0
-	}
-	if trust < 0.0 {
-		trust = 0.0
-	}
-
-	return trust
-}
-
-func (tie *TenantIsolationEngine) determineZeroTrustAccess(profile *ZeroTrustProfile, trustScore float64, request *ZeroTrustAccessRequest) bool {
-	// Access granted if trust score is above threshold
-	threshold := 0.6
-
-	// Adjust threshold based on resource sensitivity
-	if request.ResourceType == "sensitive" {
-		threshold = 0.8
-	}
-
-	return trustScore >= threshold
-}
-
-func (tie *TenantIsolationEngine) determineRequiredAuthentication(profile *ZeroTrustProfile, trustScore float64) []AuthenticationFactor {
-	if trustScore > 0.8 {
-		return []AuthenticationFactor{AuthFactorPassword}
-	} else if trustScore > 0.6 {
-		return []AuthenticationFactor{AuthFactorPassword, AuthFactorTOTP}
-	} else {
-		return []AuthenticationFactor{AuthFactorPassword, AuthFactorTOTP, AuthFactorBiometric}
-	}
-}
-
-func (tie *TenantIsolationEngine) executeAdaptationRule(ctx context.Context, tenantID string, rule AdaptationRule) error {
-	switch rule.Action {
-	case AdaptationActionTightenControls:
-		return tie.tightenSecurityControls(ctx, tenantID)
-	case AdaptationActionRequireMFA:
-		return tie.enableAdditionalMFA(ctx, tenantID, rule.Parameters)
-	case AdaptationActionLimitAccess:
-		return tie.limitTenantAccess(ctx, tenantID, rule.Parameters)
-	case AdaptationActionIncreaseMonitoring:
-		return tie.increaseMonitoring(ctx, tenantID)
-	case AdaptationActionIsolateSession:
-		return tie.isolateActiveSessions(ctx, tenantID)
-	case AdaptationActionEscalateAlert:
-		return tie.escalateSecurityAlert(ctx, tenantID, rule.Parameters)
-	}
-	return nil
-}
-
-func (tie *TenantIsolationEngine) tightenSecurityControls(ctx context.Context, tenantID string) error {
-	// Implementation would tighten various security controls
-	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("controls-%s", tenantID), "tighten_controls")
-}
-
-func (tie *TenantIsolationEngine) enableAdditionalMFA(ctx context.Context, tenantID string, params map[string]interface{}) error {
-	// Implementation would enable additional MFA requirements
-	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("mfa-%s", tenantID), "enable_mfa")
-}
-
-func (tie *TenantIsolationEngine) limitTenantAccess(ctx context.Context, tenantID string, params map[string]interface{}) error {
-	// Implementation would limit access permissions
-	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("access-%s", tenantID), "limit_access")
-}
-
-func (tie *TenantIsolationEngine) increaseMonitoring(ctx context.Context, tenantID string) error {
-	// Implementation would increase monitoring levels
-	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("monitor-%s", tenantID), "increase_monitoring")
-}
-
-func (tie *TenantIsolationEngine) isolateActiveSessions(ctx context.Context, tenantID string) error {
-	// Implementation would isolate active sessions
-	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("session-%s", tenantID), "isolate_sessions")
-}
-
-func (tie *TenantIsolationEngine) escalateSecurityAlert(ctx context.Context, tenantID string, params map[string]interface{}) error {
-	// Implementation would escalate security alerts
-	return tie.auditLogger.LogRemediationAction(ctx, fmt.Sprintf("alert-%s", tenantID), "escalate_alert")
-}
-
-// Supporting types for zero-trust access evaluation
-
-// ZeroTrustAccessEvent represents an access event for behavioral analysis
-type ZeroTrustAccessEvent struct {
-	EventID    string                 `json:"event_id"`
-	TenantID   string                 `json:"tenant_id"`
-	UserID     string                 `json:"user_id"`
-	DeviceID   string                 `json:"device_id"`
-	ResourceID string                 `json:"resource_id"`
-	Action     string                 `json:"action"`
-	Timestamp  time.Time              `json:"timestamp"`
-	Location   string                 `json:"location,omitempty"`
-	Success    bool                   `json:"success"`
-	Duration   time.Duration          `json:"duration,omitempty"`
-	DataVolume float64                `json:"data_volume,omitempty"`
-	Context    map[string]interface{} `json:"context,omitempty"`
-}
-
-// ZeroTrustAccessRequest represents a zero-trust access request
-type ZeroTrustAccessRequest struct {
-	RequestID    string                 `json:"request_id"`
-	TenantID     string                 `json:"tenant_id"`
-	UserID       string                 `json:"user_id"`
-	DeviceID     string                 `json:"device_id"`
-	ResourceID   string                 `json:"resource_id"`
-	ResourceType string                 `json:"resource_type"`
-	Action       string                 `json:"action"`
-	Timestamp    time.Time              `json:"timestamp"`
-	Location     string                 `json:"location,omitempty"`
-	Context      map[string]interface{} `json:"context,omitempty"`
-}
-
-// ZeroTrustAccessResponse represents the response to a zero-trust access request
-type ZeroTrustAccessResponse struct {
-	RequestID    string                 `json:"request_id"`
-	TenantID     string                 `json:"tenant_id"`
-	Granted      bool                   `json:"granted"`
-	TrustLevel   ZeroTrustLevel         `json:"trust_level"`
-	RiskScore    float64                `json:"risk_score"`
-	Reason       string                 `json:"reason"`
-	RequiredAuth []AuthenticationFactor `json:"required_auth"`
-	EvaluatedAt  time.Time              `json:"evaluated_at"`
-	Constraints  []string               `json:"constraints,omitempty"`
-	SessionData  map[string]interface{} `json:"session_data,omitempty"`
 }
