+++ features/tenant/security/vulnerability_manager.go
@@ -1,0 +1,439 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Jordan Ritz
+package security
+
+import (
+	"context"
+	"fmt"
+	"time"
+)
+
+// VulnerabilityType defines types of security vulnerabilities
+type VulnerabilityType string
+
+const (
+	VulnerabilityMissingPatches      VulnerabilityType = "missing_patches"
+	VulnerabilityWeakCredentials     VulnerabilityType = "weak_credentials"
+	VulnerabilityOpenPorts           VulnerabilityType = "open_ports"
+	VulnerabilityOutdatedSoftware    VulnerabilityType = "outdated_software"
+	VulnerabilityMisconfiguration    VulnerabilityType = "misconfiguration"
+	VulnerabilityPrivilegeEscalation VulnerabilityType = "privilege_escalation"
+	VulnerabilityDataExposure        VulnerabilityType = "data_exposure"
+)
+
+// VulnerabilitySeverity defines vulnerability severity levels
+type VulnerabilitySeverity string
+
+const (
+	VulnerabilitySeverityLow      VulnerabilitySeverity = "low"
+	VulnerabilitySeverityMedium   VulnerabilitySeverity = "medium"
+	VulnerabilitySeverityHigh     VulnerabilitySeverity = "high"
+	VulnerabilitySeverityCritical VulnerabilitySeverity = "critical"
+)
+
+// RemediationAction defines automated remediation actions
+type RemediationAction string
+
+const (
+	RemediationActionQuarantine    RemediationAction = "quarantine"
+	RemediationActionPatch         RemediationAction = "patch"
+	RemediationActionDisable       RemediationAction = "disable"
+	RemediationActionAlert         RemediationAction = "alert"
+	RemediationActionBlock         RemediationAction = "block"
+	RemediationActionRotateSecrets RemediationAction = "rotate_secrets"
+	RemediationActionIsolate       RemediationAction = "isolate"
+)
+
+// Vulnerability represents a detected security vulnerability
+type Vulnerability struct {
+	ID            string                `json:"id"`
+	TenantID      string                `json:"tenant_id"`
+	Type          VulnerabilityType     `json:"type"`
+	Severity      VulnerabilitySeverity `json:"severity"`
+	Title         string                `json:"title"`
+	Description   string                `json:"description"`
+	AffectedAsset string                `json:"affected_asset"`
+	CVEID         string                `json:"cve_id,omitempty"`
+	CVSSScore     float64               `json:"cvss_score,omitempty"`
+	DetectedAt    time.Time             `json:"detected_at"`
+	UpdatedAt     time.Time             `json:"updated_at"`
+	Status        string                `json:"status"` // "open", "remediated", "mitigated", "accepted"
+	Remediation   *RemediationPlan      `json:"remediation,omitempty"`
+}
+
+// RemediationPlan defines how to remediate a vulnerability
+type RemediationPlan struct {
+	Actions       []RemediationAction `json:"actions"`
+	AutoRemediate bool                `json:"auto_remediate"`
+	ManualSteps   []string            `json:"manual_steps"`
+	EstimatedTime time.Duration       `json:"estimated_time"`
+	Priority      int                 `json:"priority"` // 1-10, where 1 is highest priority
+	Dependencies  []string            `json:"dependencies"`
+	RollbackPlan  []string            `json:"rollback_plan"`
+	CreatedAt     time.Time           `json:"created_at"`
+	ScheduledAt   *time.Time          `json:"scheduled_at,omitempty"`
+	CompletedAt   *time.Time          `json:"completed_at,omitempty"`
+}
+
+// VulnerabilityAssessment contains the result of vulnerability scanning
+type VulnerabilityAssessment struct {
+	TenantID         string          `json:"tenant_id"`
+	ScanID           string          `json:"scan_id"`
+	ScanTime         time.Time       `json:"scan_time"`
+	Vulnerabilities  []Vulnerability `json:"vulnerabilities"`
+	RiskScore        float64         `json:"risk_score"`
+	ComplianceStatus map[string]bool `json:"compliance_status"`
+	Recommendations  []string        `json:"recommendations"`
+}
+
+// ScanForVulnerabilities performs a comprehensive vulnerability assessment for a tenant
+func (tie *TenantIsolationEngine) ScanForVulnerabilities(ctx context.Context, tenantID string) (*VulnerabilityAssessment, error) {
+	tie.mutex.Lock()
+	defer tie.mutex.Unlock()
+
+	// Validate tenant exists
+	_, err := tie.tenantManager.GetTenant(ctx, tenantID)
+	if err != nil {
+		return nil, fmt.Errorf("tenant not found: %w", err)
+	}
+
+	assessment := &VulnerabilityAssessment{
+		TenantID:         tenantID,
+		ScanID:           fmt.Sprintf("scan-%s-%d", tenantID, time.Now().Unix()),
+		ScanTime:         time.Now(),
+		Vulnerabilities:  []Vulnerability{},
+		ComplianceStatus: make(map[string]bool),
+		Recommendations:  []string{},
+	}
+
+	// Get isolation rule for compliance requirements
+	rule, _ := tie.GetIsolationRule(ctx, tenantID)
+
+	// Simulate vulnerability detection (in production, this would integrate with actual scanners)
+	vulnerabilities := tie.detectVulnerabilities(tenantID, rule)
+	assessment.Vulnerabilities = vulnerabilities
+
+	// Calculate risk score based on vulnerabilities
+	assessment.RiskScore = tie.calculateRiskScore(vulnerabilities)
+
+	// Check compliance status
+	assessment.ComplianceStatus = tie.checkComplianceStatus(rule, vulnerabilities)
+
+	// Generate recommendations
+	assessment.Recommendations = tie.generateRecommendations(vulnerabilities, rule)
+
+	// Store vulnerabilities for tracking
+	tie.vulnerabilities[tenantID] = vulnerabilities
+
+	// Auto-create remediation plans for critical vulnerabilities
+	for _, vuln := range vulnerabilities {
+		if vuln.Severity == VulnerabilitySeverityCritical || vuln.Severity == VulnerabilitySeverityHigh {
+			plan := tie.createRemediationPlan(vuln, rule)
+			tie.remediationPlans[vuln.ID] = plan
+		}
+	}
+
+	return assessment, nil
+}
+
+// detectVulnerabilities simulates vulnerability detection
+func (tie *TenantIsolationEngine) detectVulnerabilities(tenantID string, rule *IsolationRule) []Vulnerability {
+	vulnerabilities := []Vulnerability{}
+	now := time.Now()
+
+	// Check for weak credential policies
+	if !rule.NetworkIsolation.RequireMTLS {
+		vulnerabilities = append(vulnerabilities, Vulnerability{
+			ID:            fmt.Sprintf("vuln-%s-weak-auth", tenantID),
+			TenantID:      tenantID,
+			Type:          VulnerabilityWeakCredentials,
+			Severity:      VulnerabilitySeverityHigh,
+			Title:         "Weak Authentication Configuration",
+			Description:   "Mutual TLS is not required, potentially allowing weaker authentication methods",
+			AffectedAsset: "authentication_system",
+			DetectedAt:    now,
+			UpdatedAt:     now,
+			Status:        "open",
+		})
+	}
+
+	// Check for inadequate encryption
+	if !rule.DataResidency.RequireEncryption || rule.DataResidency.EncryptionLevel == "standard" {
+		severity := VulnerabilitySeverityMedium
+		if !rule.DataResidency.RequireEncryption {
+			severity = VulnerabilitySeverityHigh
+		}
+		vulnerabilities = append(vulnerabilities, Vulnerability{
+			ID:            fmt.Sprintf("vuln-%s-encryption", tenantID),
+			TenantID:      tenantID,
+			Type:          VulnerabilityDataExposure,
+			Severity:      severity,
+			Title:         "Inadequate Data Encryption",
+			Description:   "Data encryption requirements are insufficient for security compliance",
+			AffectedAsset: "data_storage",
+			DetectedAt:    now,
+			UpdatedAt:     now,
+			Status:        "open",
+		})
+	}
+
+	// Check for overly permissive cross-tenant access
+	if rule.CrossTenantAccess.AllowCrossTenantAccess && !rule.CrossTenantAccess.RequireApproval {
+		vulnerabilities = append(vulnerabilities, Vulnerability{
+			ID:            fmt.Sprintf("vuln-%s-cross-tenant", tenantID),
+			TenantID:      tenantID,
+			Type:          VulnerabilityPrivilegeEscalation,
+			Severity:      VulnerabilitySeverityMedium,
+			Title:         "Overly Permissive Cross-Tenant Access",
+			Description:   "Cross-tenant access is allowed without requiring approval, increasing attack surface",
+			AffectedAsset: "access_control",
+			DetectedAt:    now,
+			UpdatedAt:     now,
+			Status:        "open",
+		})
+	}
+
+	// Check for insufficient resource isolation
+	if !rule.ResourceIsolation.IsolatedStorage {
+		vulnerabilities = append(vulnerabilities, Vulnerability{
+			ID:            fmt.Sprintf("vuln-%s-isolation", tenantID),
+			TenantID:      tenantID,
+			Type:          VulnerabilityMisconfiguration,
+			Severity:      VulnerabilitySeverityHigh,
+			Title:         "Insufficient Resource Isolation",
+			Description:   "Tenant resources are not properly isolated, potentially allowing data leakage",
+			AffectedAsset: "resource_isolation",
+			DetectedAt:    now,
+			UpdatedAt:     now,
+			Status:        "open",
+		})
+	}
+
+	return vulnerabilities
+}
+
+// calculateRiskScore calculates an overall risk score based on vulnerabilities
+func (tie *TenantIsolationEngine) calculateRiskScore(vulnerabilities []Vulnerability) float64 {
+	if len(vulnerabilities) == 0 {
+		return 0.0
+	}
+
+	var totalScore float64
+	weights := map[VulnerabilitySeverity]float64{
+		VulnerabilitySeverityLow:      1.0,
+		VulnerabilitySeverityMedium:   3.0,
+		VulnerabilitySeverityHigh:     7.0,
+		VulnerabilitySeverityCritical: 10.0,
+	}
+
+	for _, vuln := range vulnerabilities {
+		if weight, exists := weights[vuln.Severity]; exists {
+			totalScore += weight
+		}
+	}
+
+	// Normalize to 0-100 scale
+	maxPossibleScore := float64(len(vulnerabilities)) * 10.0
+	if maxPossibleScore > 0 {
+		return (totalScore / maxPossibleScore) * 100.0
+	}
+
+	return 0.0
+}
+
+// checkComplianceStatus checks compliance against various standards
+func (tie *TenantIsolationEngine) checkComplianceStatus(rule *IsolationRule, vulnerabilities []Vulnerability) map[string]bool {
+	status := make(map[string]bool)
+
+	// Check basic compliance requirements
+	status["basic"] = true
+	for _, vuln := range vulnerabilities {
+		if vuln.Severity == VulnerabilitySeverityCritical {
+			status["basic"] = false
+			break
+		}
+	}
+
+	// Check specific compliance standards based on tenant's compliance level
+	switch rule.ComplianceLevel {
+	case ComplianceLevelHIPAA:
+		status["hipaa"] = rule.DataResidency.RequireEncryption &&
+			rule.DataResidency.EncryptionLevel != "standard" &&
+			rule.ResourceIsolation.IsolatedStorage
+	case ComplianceLevelPCIDSS:
+		status["pci_dss"] = rule.NetworkIsolation.RequireMTLS &&
+			rule.DataResidency.RequireEncryption &&
+			rule.ResourceIsolation.IsolatedStorage
+	case ComplianceLevelFedRAMP:
+		status["fedramp"] = rule.DataResidency.EncryptionLevel == "fips" &&
+			rule.NetworkIsolation.RequireMTLS &&
+			!rule.CrossTenantAccess.AllowCrossTenantAccess
+	}
+
+	return status
+}
+
+// generateRecommendations generates security recommendations based on vulnerabilities
+func (tie *TenantIsolationEngine) generateRecommendations(vulnerabilities []Vulnerability, rule *IsolationRule) []string {
+	recommendations := []string{}
+
+	for _, vuln := range vulnerabilities {
+		switch vuln.Type {
+		case VulnerabilityWeakCredentials:
+			recommendations = append(recommendations, "Enable mutual TLS authentication")
+			recommendations = append(recommendations, "Implement multi-factor authentication")
+		case VulnerabilityDataExposure:
+			recommendations = append(recommendations, "Upgrade to high-level encryption")
+			recommendations = append(recommendations, "Enable encryption at rest and in transit")
+		case VulnerabilityPrivilegeEscalation:
+			recommendations = append(recommendations, "Require approval for cross-tenant access")
+			recommendations = append(recommendations, "Implement least privilege access controls")
+		case VulnerabilityMisconfiguration:
+			recommendations = append(recommendations, "Enable isolated storage for tenant resources")
+			recommendations = append(recommendations, "Review and update security configurations")
+		}
+	}
+
+	return recommendations
+}
+
+// createRemediationPlan creates an automated remediation plan for a vulnerability
+func (tie *TenantIsolationEngine) createRemediationPlan(vuln Vulnerability, rule *IsolationRule) *RemediationPlan {
+	plan := &RemediationPlan{
+		Actions:       []RemediationAction{},
+		AutoRemediate: false,
+		ManualSteps:   []string{},
+		EstimatedTime: time.Hour,
+		Priority:      5,
+		Dependencies:  []string{},
+		RollbackPlan:  []string{},
+		CreatedAt:     time.Now(),
+	}
+
+	switch vuln.Type {
+	case VulnerabilityWeakCredentials:
+		plan.Actions = []RemediationAction{RemediationActionAlert}
+		plan.ManualSteps = []string{
+			"Enable mutual TLS in network isolation settings",
+			"Update authentication configuration",
+			"Test authentication with new settings",
+		}
+		plan.Priority = 2
+		plan.EstimatedTime = time.Hour * 2
+
+	case VulnerabilityDataExposure:
+		plan.Actions = []RemediationAction{RemediationActionAlert}
+		plan.ManualSteps = []string{
+			"Upgrade encryption level to 'high' or 'fips'",
+			"Enable encryption requirements",
+			"Verify data encryption across all systems",
+		}
+		plan.Priority = 1
+		plan.EstimatedTime = time.Hour * 4
+
+	case VulnerabilityPrivilegeEscalation:
+		plan.Actions = []RemediationAction{RemediationActionAlert}
+		plan.ManualSteps = []string{
+			"Enable approval requirement for cross-tenant access",
+			"Review existing cross-tenant permissions",
+			"Update access control policies",
+		}
+		plan.Priority = 3
+		plan.EstimatedTime = time.Hour * 1
+
+	case VulnerabilityMisconfiguration:
+		plan.Actions = []RemediationAction{RemediationActionAlert}
+		plan.ManualSteps = []string{
+			"Enable isolated storage in resource isolation settings",
+			"Migrate shared resources to isolated storage",
+			"Verify resource isolation boundaries",
+		}
+		plan.Priority = 2
+		plan.EstimatedTime = time.Hour * 6
+	}
+
+	if vuln.Severity == VulnerabilitySeverityCritical {
+		plan.Priority = 1
+		plan.Actions = append(plan.Actions, RemediationActionIsolate)
+	}
+
+	return plan
+}
+
+// GetVulnerabilities returns vulnerabilities for a tenant
+func (tie *TenantIsolationEngine) GetVulnerabilities(ctx context.Context, tenantID string) ([]Vulnerability, error) {
+	tie.mutex.RLock()
+	defer tie.mutex.RUnlock()
+
+	vulnerabilities, exists := tie.vulnerabilities[tenantID]
+	if !exists {
+		return []Vulnerability{}, nil
+	}
+
+	return vulnerabilities, nil
+}
+
+// UpdateVulnerabilityStatus updates the status of a vulnerability
+func (tie *TenantIsolationEngine) UpdateVulnerabilityStatus(ctx context.Context, vulnerabilityID, status string) error {
+	tie.mutex.Lock()
+	defer tie.mutex.Unlock()
+
+	// Find and update the vulnerability
+	for tenantID, vulnerabilities := range tie.vulnerabilities {
+		for i, vuln := range vulnerabilities {
+			if vuln.ID == vulnerabilityID {
+				vulnerabilities[i].Status = status
+				vulnerabilities[i].UpdatedAt = time.Now()
+
+				// If remediated, update remediation plan
+				if status == "remediated" {
+					if plan, exists := tie.remediationPlans[vulnerabilityID]; exists {
+						now := time.Now()
+						plan.CompletedAt = &now
+					}
+				}
+
+				// Audit the status change
+				return tie.auditLogger.LogVulnerabilityStatusChange(ctx, vulnerabilityID, tenantID, status)
+			}
+		}
+	}
+
+	return fmt.Errorf("vulnerability not found: %s", vulnerabilityID)
+}
+
+// ExecuteRemediationPlan executes an automated remediation plan
+func (tie *TenantIsolationEngine) ExecuteRemediationPlan(ctx context.Context, vulnerabilityID string) error {
+	tie.mutex.Lock()
+	defer tie.mutex.Unlock()
+
+	plan, exists := tie.remediationPlans[vulnerabilityID]
+	if !exists {
+		return fmt.Errorf("remediation plan not found for vulnerability: %s", vulnerabilityID)
+	}
+
+	if !plan.AutoRemediate {
+		return fmt.Errorf("remediation plan requires manual execution")
+	}
+
+	// Execute automated actions
+	for _, action := range plan.Actions {
+		switch action {
+		case RemediationActionAlert:
+			// Send alert to administrators
+			_ = tie.auditLogger.LogRemediationAction(ctx, vulnerabilityID, string(action))
+		case RemediationActionIsolate:
+			// Temporarily isolate affected resources
+			_ = tie.auditLogger.LogRemediationAction(ctx, vulnerabilityID, string(action))
+		case RemediationActionBlock:
+			// Block suspicious access
+			_ = tie.auditLogger.LogRemediationAction(ctx, vulnerabilityID, string(action))
+		}
+	}
+
+	// Mark as scheduled for execution
+	now := time.Now()
+	plan.ScheduledAt = &now
+
+	return nil
+}
